Took me a couple of hours as I had to choose a workable tool to generate the hashes that this required.
I can take no credit for the real work here.

This is just a workup of what could be done to address potential exploits
that modify system, or other critical files. This provides a means to monitor
the state of critical files; guarding against modification, and providing
a mechanism to generate meaningful preemptive security alerts .

Amiel.